Move ExtensionLoadingMaster
